+++ src/main/java/com/robertx22/age_of_exile/loot/blueprints/bases/SkillGemTypePart.java
@@ -13,21 +13,8 @@ public class SkillGemTypePart extends BlueprintPart<SkillGem, SkillGemBlueprint>
     @Override
     protected SkillGem generateIfNull() {
 
-        try {
-
-            if (blueprint.info != null) {
-                if (blueprint.info.playerData == null && blueprint.info.level < 20) {
-                    return ExileDB.SkillGems()
-                        .getFilterWrapped(x -> x.req.meetsReq(blueprint.info.level, blueprint.info.playerData))
-                        .random();
-                }
-
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-
         return ExileDB.SkillGems()
             .random();
+
     }
 }