+++ CHANGELOG.md
@@ -13,4 +13,5 @@
 - gear stat requirements removed
 - removed legendary and mythic rarity
 - runeword gear is now crafted in runeword station by combining all the runes needed
-- ramoved magus mobs
+- removed magus mobs
+- removed spell action requirements